+++ sqlstore/accounts_test.go
@@ -14,8 +14,8 @@ func addTestAccount(t *testing.T,
 	accountStore *sqlstore.Accounts,
 	party entities.Party,
 	asset entities.Asset,
-	block entities.Block) entities.Account {
-
+	block entities.Block,
+) entities.Account {
 	account := entities.Account{
 		PartyID:  party.ID,
 		AssetID:  asset.ID,
@@ -85,5 +85,4 @@ func TestAccount(t *testing.T) {
 	accs, err = accountStore.Query(filter)
 	assert.NoError(t, err)
 	assert.Len(t, accs, 0)
-
 }