+++ sqlstore/migrations/0001_initial.sql
@@ -99,6 +99,34 @@ CREATE VIEW orders_current_versions AS (
   SELECT DISTINCT ON (id, version) * FROM orders ORDER BY id, version DESC, vega_time DESC
 );
 
+create table trades
+(
+    vega_time       TIMESTAMP WITH TIME ZONE NOT NULL REFERENCES blocks(vega_time),
+    seq_num    BIGINT NOT NULL,
+    id     BYTEA NOT NULL,
+    market_id BYTEA NOT NULL,
+    price     NUMERIC(32, 0),
+    size      NUMERIC(32, 0),
+    buyer     BYTEA NOT NULL,
+    seller    BYTEA NOT NULL,
+    aggressor SMALLINT,
+    buy_order BYTEA NOT NULL,
+    sell_order BYTEA NOT NULL,
+    type       SMALLINT NOT NULL,
+    buyer_maker_fee NUMERIC(32, 0),
+    buyer_infrastructure_fee NUMERIC(32, 0),
+    buyer_liquidity_fee NUMERIC(32, 0),
+    seller_maker_fee NUMERIC(32, 0),
+    seller_infrastructure_fee NUMERIC(32, 0),
+    seller_liquidity_fee NUMERIC(32, 0),
+    buyer_auction_batch BIGINT,
+    seller_auction_batch BIGINT
+);
+
+SELECT create_hypertable('trades', 'vega_time', chunk_time_interval => INTERVAL '1 day');
+CREATE INDEX ON trades (market_id, vega_time DESC);
+
+
 CREATE TABLE network_limits (
   vega_time                   TIMESTAMP WITH TIME ZONE NOT NULL PRIMARY KEY REFERENCES blocks(vega_time),
   can_propose_market          BOOLEAN NOT NULL,
@@ -147,7 +175,7 @@ create type market_trading_mode_type as enum('TRADING_MODE_UNSPECIFIED', 'TRADIN
 create table market_data (
     market bytea not null,
     vega_time timestamp with time zone not null references blocks(vega_time),
-    seq_num int not null,
+    seq_num bigint not null,
     mark_price numeric(32),
     best_bid_price numeric(32),
     best_bid_volume bigint,
@@ -223,5 +251,6 @@ DROP TABLE IF EXISTS balances;
 DROP TABLE IF EXISTS accounts;
 DROP TABLE IF EXISTS parties;
 DROP TABLE IF EXISTS assets;
+DROP TABLE IF EXISTS trades;
 DROP TABLE IF EXISTS blocks;
 DROP EXTENSION IF EXISTS timescaledb;